workflow_files: move clean functionality into its own module
* `cylc clean` has outgrown the workflow_files module somewhat.
